Count capability usage by effective Agent version
